LayoutTests/media/document-level-media-user-gesture-quirk-expected.txt

  
RUN(internals.setMediaElementRestrictions(video1, "RequireUserGestureForAudioRateChange"))
RUN(internals.setMediaElementRestrictions(video2, "RequireUserGestureForAudioRateChange"))
RUN(internals.settings.setNeedsSiteSpecificQuirks(true))
Trying to play the first video without a user gesture
RUN(video1.play().then(failTest).catch(playFirstVideoWithUserGesture))
Playback was denied - trying to play the first video with a user gesture
RUN(video1.play().then(playedFirstVideo).catch(failTest))
The first video played with a user gesture - trying to play the second video
The second video should play unconditionally on Mac with quirks enabled
RUN(video2.play().then(endTest).catch(failTest))
END OF TEST
